remove dev menu and move user timings
