+++ shared/util/user-timings.js
